Inspector v2: Improve properties pane perf (#17085)
Previously there was a kind of hacky thing being done where when the
property pane is bound to a new entity, it sets the Accordion key to
some new unique value, which causes React to re-create the Accordion
component (not just re-render it), which is done to get the
defaultOpenItems (accordion sections) to be re-evaluated. But this also
leads to a perf problem, because Accordion does some heavy work when it
is first mounted, so continually re-mounting a new instance (as opposed
to just re-rendering the same component instance) is expensive and adds
to the noticeable delay when a new entity is selected.

This change is makes the logic for dealing with accordion section open
state smarter, so we don't have to keep re-creating the Accordion.

I also added a little more memo-ization to avoid unnecessary re-renders
of the property pane, as well as adding displayNames to some of the
components so they don't show up as "Anonymous" in the React dev tools
profiler. @georginahalpern we should add displayName to most of our
shared components, particularly the ones that use forwardRef and are
otherwise unnamed.