Inspector v2: Setup public package (#17086)
This PR includes the local build setup for the Inspector v2 package. It
follows the same pattern as the [Viewer v2 package
](https://github.com/BabylonJS/Babylon.js/pull/15246/) before it fully
replaced Viewer v1.

- Like the Viewer, this uses Rollup to create an esm bundle. This is
needed right now because we don't have a published package for
shared-ui-components, so it needs to be bundled in. Once we have a
published package for the shared ui components (e.g.
@babylonjs/tools-core or something), then we could remove Rollup and
just compile the source. NOTE: I spent a couple days trying to make this
work with WebPack, but I really don't think it is possible. WebPack is
just not setup to be able to create esm bundles.
- Within the mono repo, the package identity is @babylonjs/inspector-v2.
However, in the build pipeline we will run the prepublish.cjs script
before the publishing step, which sets up the inspector-v2 package with
the identity @babylonjs/inspector, but with `-preview` appended to the
version number (and we will add the @preview tag as well), sets the
correct version, and removes the private flag (so it can be published to
npmjs).
- Also had to fix a bunch of places in shared-ui-components and
inspector-v2 code where we accidentally had imports from index files.