Add BVH exporter (#17093)
Adds a BVH exporter that can generate a BVH file from a skeleton.

You can pass the names of animations you want to export and they are
appended to the same motion data.

If the frame rate is not set, it uses 0.333 (30 FPS) by default.

The bvh fixture file was acquired from the repository:
https://github.com/BandaiNamcoResearchInc/
License: *CC-BY-NC*