Smart Filters Optimizer: disallow uniform access from helpers (#17105)
To keep the resulting shaders small, the optimizer tries to emit only
one instance of a helper function (any function that's not the main
function of the shader) regardless of the number of instances of that
block are being combined. Previously it would detect if a uniform was
being accessed and wouldn't combine instance in that case, but there is
a case where main() calls foo() calls bar() where foo() didn't access a
uniform, so it was consolidated, while bar() did, so it could not be. To
prevent this bug, this change adds a clear error message if a helper
function accesses a uniform.