Inspector v2: Fix crash when showing scene properties (#17118)
If a PhysicsEngine is not imported with side effects setting things up,
then `scene.getPhysicsEngine` returns `undefined`, even though according
to the type it should return `null`. Additionally, the `SceneProperties`
for Inspector v2 only checks for a `null` physics engine, not
`undefined`, which leads to an unhandled error. This PR fixes both of
these - `getPhysicsEngine` returns `null` rather than `undefined`
(consistent with the declared type) and `SceneProperties` is more
defensive and checks for both `null` and `undefined`.

@PatrickRyanMS FYI