Adding visual tests using the devhost (#17124)
This PR enables visual tests for projects that use the devhost instead
of the babylon server.

For this:
- npm run start:devhost -> Will start the devhost for testing
- npm run test:visualization:devhost -> Will run all visualization tests
that use the devhost

All other existing npm commands stay as is.

playwright.config.ts got updated, so now there are:
- playwright.config.ts -> for tests that use the babylon server
- playwright.devhost.config.ts -> for tests that use the devhost
- playwright.utils.ts -> shared code for all playground tests, and
contains the list of tests for each test environment

And as an example case, 20 visual tests are added for Lottie, which use
the devhost. These tests follow the same structure as existing visual
tests:
- visualization.lottie.test.ts -> gets the list of files to test for
lottie
- config.lottie.json -> the definition of the tests