Inspector v2: Remove all usages of constructor.name (#17126)
This change removes all usages of `constructor.name`. Conceptually it's
a nice idea, but unfortunately it doesn't work when the code is minified
because the class names become arbitrary short (often one character)
names that are completely unrelated to the "real" original class names.
I only noticed this issue in the deployed Playground where the code is
minified (it's not minified in local dev builds).