ESLint: Fix a couple issues in the cross package relative import custom rule (#17127)
This fixes a couple of issues:
1. There was a hardcoded relative path I missed in the original PR,
which means the rule wasn't working correctly for files deeper in the
directory structure (failed testing!). Fixed by re-using the correctly
calculated project root (which already existed and just needed to be
re-used in another part of the code).
2. Once the first issue was fixed, it also revealed a second bug where
we were not always correctly recognizing when a relative path was within
the same package/project, and incorrectly flagging those imports. Fixed
by getting the tsconfig path mapping for both the file being analyzed
and the import, and checking if they are from different
projects/packages (or at least different path mappings, which is
basically the same).