Atmosphere add-on (#17107)
Provides an add-on for rendering a physically based atmosphere,
supporting aerial perspective, sky, and views from ground to space.
Playground ID: **#EQ2D4A#1**

This implementation is based on __A Scalable and Production Ready Sky
and Atmosphere Rendering Technique__, Hillaire, S. (2020), and
https://github.com/sebh/UnrealEngineSkyAtmosphere

Note, the new atmosphere add-on is intended to work well with
PBRMaterials. When the atmosphere is constructed, a MaterialPlugin is
activated which enables PBRMaterials to use atmospheric effects,
modifying the direct light contribution as well as the irradiance term
in the PBR shader.

A custom inspector has also been provided to make it easier to modify
the atmosphere parameters. (Default settings represent an Earth-like
atmosphere.)