Fix legacy audio engine dispose issue (#17150)
The v2 audio engine's main out node is being reset to the legacy audio
engine's master gain node but the v2 main bus node is not being informed
of the change, so it fails in `disconnect` and throws an error.

This change fixes the issue by updating the internal WebAudio node
connections appropriately in the legacy audio engine.