Avoid type error when updating WebXRCamera reference space
Per the MDN docs the `XRRigidTransform` constructor's `position` argument takes a `w` coord in addition to the `x/y/z` coords, and `w` must be set to `1` otherwise a `TypeError` exception is thrown. This does not seem to be strictly enforced, but a user has reported the exception is thrown in some cases (see https://forum.babylonjs.com/t/xrrigidtransform-invalid-position-error/60416).

This change fixes the issue by always setting the `w` coordinate to `1`.

Tested with playground https://playground.babylonjs.com/#AZML8U#261.