Smart Filter Debugger: Fix CSS issue when host has custom body tag styles, fix VSCode module resolution (#17162)
The way popup work in our shared code, the CSS styles for the contents
of the popup get added to the original window, then copied into the
popup. In the case of the SF debugger browser extension, this can copy
in CSS from the hosting page that we don't want - in this case, it was
custom styles on the body tag. This is a small fix to ensure the popup
for the SF editor always has the correct display style on the body tag,
regardless of the value from the hosting page.

I also noticed that the tsconfig.json used by VS was incorrectly
referencing the root tsconfig.build.json, so F12 took you to dist
instead of src dependencies - fixed that too.