Inspector v2: Fix a couple property pane bugs (#17171)
This change fixes two issues:
1. An unhandled error that can occur when switching between nodes with
and without animations. This was happening due to a violation of the
rule of hooks - specifically, there was a conditional hook, which
completely breaks React's internal state.
2. Set the flexDirection on the `Collapse` wrapper component to match
the collapse orientation. There is no way for us to know the intent of
the caller if they use a React fragment, but defaulting to the collapse
orientation seems reasonable (and fixes the issues I was seeing in the
property pane). If a user wants to override the default, they can just
use a div explicitly rather than a React fragment.