OBJFileLoader: Update defaults for missing mesh attributes (#17168)
This PR updates how missing mesh attributes are handled.
- UVs: Previously, missing UVs were replaced with a default value of (0,
0) in the buffer. Now, the UV buffer will remain undefined if UVs are
not present.
- Normals: Previously, missing normals were replaced with a default
value of (0, 1, 0) in the buffer. Now, normals will be computed at the
end for smooth shading.

This is a breaking change for users expecting the mesh's `UVKind` vertex
buffer to be defined.

<details>
<summary>Misc</summary>

- Check whether colors are present instead whether `importVertexColor`
is true (which will still be the case)
- Use object param for _setData
</details>