Disable chunking for umd builds (#17178)
This disables chunking for umd builds as we always want one chunk per
entry point (usually we just have one entry point, but not always, such
as with the loaders package). Without this, various circumstances can
lead to chunking (such as multiple entry points with a shared dependency
on an internal only API).