Inspector v2: Add exports for inspector and shared ui components needed for extending inspector features (#17174)
This PR is mostly about adding the needed exports to extend the
inspector with new features. This includes:
1. Exports from the inspector package itself - these I did selectively
from the root index.ts, because what should be part of the public API is
only a subset of what is used internally within the inspector itself and
future internal tool usage.
2. Exports from the shared ui package - these I did with an index.ts
hierarchy as the intention is that 95% of the Fluent directory should be
exported, so I thought this made more sense for this case.

I also included a small change in this PR where Scene Explorer now hides
empty sections (e.g. if there are no post proesses, the "Post Processes"
section won't be displayed). This is to minimize the noise/cognitive
load when you are only using a subset of the Babylon features.