Create abstraction layer over the inspector gizmo layer (#17177)
This change fixes some issues with the inspector due to its
`gizmoManager` being treated as an `any` type.

## Issue Description

To see the issue this is attempting to address, go to
[https://sandbox.babylonjs.com/](https://sandbox.babylonjs.com/?asset=https://playground.babylonjs.com/scenes/skull.babylon)
and load a model, then open the inspector and click this button:
<img width="536" height="220" alt="image"
src="https://github.com/user-attachments/assets/638ce37e-f310-40f8-a424-e557f35ab062"
/>

A message will appear on screen: `Script error. Check the developer
console.` (nothing appears in console).

The actual issue stems from this error:
```log
sceneTreeItemComponent.tsx:185  Uncaught TypeError: Cannot read properties of null (reading 'gizmoManager')
// or
sceneTreeItemComponent.tsx:187  Uncaught TypeError: Cannot set properties of null (setting 'coordinatesMode')
```

Which leads to here:

https://github.com/BabylonJS/Babylon.js/blob/197d8f6728cc57e46df8e63eaab776a023e5a3ff/packages/dev/inspector/src/components/sceneExplorer/entities/sceneTreeItemComponent.tsx#L185-L187

The issue being that this assumes that this value has been initialized,
which in this case it has not been.

## Change Description

This change makes all access to the Inspector's `gizmoManager` use one
of two helper functions: one for getting and optionally creating, and
one for disposing. This abstracts the type-unsafety of the `any` types
to one file, rather than spreading them across ~10 different locations.

This code should not make any changes to when the inspector creates the
`gizmoManager`, as the code path to create it should only be activated
in the place where it was originally created.