Add floatingOriginMode to the scene (#17183)
This PR enables floatingOriginMode on the scene (defined at scene
construction time), which will keep track of the active camera's
position and use that as the floatingOriginOffset when sending uniform
values to shader.

In this mode we center the camera at world origin and offset all meshes
by the floatingOriginOffset. This helps avoid floating point imprecision
issues when dealing with large real world coordinates.

This behavior is achieved by overriding the uniformBuffer / effect
setMatrix functions to set viewMatrix translation == (0,0,0),
worldMatrix translation += floatingOriginOffset, and re-multiplying the
resulting worldView/viewProjection/worldViewProjection matrices with
offset.

We also offset the light position (stored in vLightData) as well as the
vEyePosition uniforms. This is done at the callsite (rather than
overriding at the effect/uniform layer).

---------

Co-authored-by: Georgina Halpern <gehalper@microsoft.com>