Geospatial Camera (#17192)
This is an experimental feature as the API will change in the next month
or so to introduce the following
- Internal understanding of the planet being orbited (radius of the
planet, coordinate system of the planet, etc)
- Latitude/Longitude/Altitude conversion atop the existing cartesian
coordinate system (this will be implemented in an extensible way so that
we are not tying the geospatial camera to earth's coordinates, rather it
can be used for other planets/worlds)
- Configurable limits to the tilt/pitch/zoom interactions
- Standard cameracentric rotation (in addition to geocentric rotation)
- Keyboard input interactions
- Fixing interactions at the poles
- And more!

Today the camera can rightClick+drag to rotate, leftClick+drag to 'drag'
globe, mousewheel to zoom to cursor or center of screen (with min/max
height constraints applied). Zoomout is in progress, you will notice
some odd interactions today.

Example playground for testing: http://localhost:1338/#7XXRHQ#11
<img width="444" height="17" alt="image"
src="https://github.com/user-attachments/assets/20897660-ffd4-4c77-915a-523cf14ce5e9"
/>

---------

Co-authored-by: Georgina Halpern <gehalper@microsoft.com>