Inspector v2: Embed mode (#17198)
This change is mostly about adding `embedMode` support in Inspector v2
(same as Inspector v1). In this mode, the panes are all pushed to the
right and stacked vertically. The way this is handled for Inspector v2
is as follows:
- Previously there was only left and right panes. Now the concept has
been expanded to allow for top left, bottom left, top right, and bottom
right. Basically split into quadrants. If there are no panes in a given
quadrant, it is not shown.
- If there are no panes on the left, and we have toolbar items in the
left toolbars (gizmo toolbar as an example), and we are in "compact"
toolbar mode (separate toolbars at the tops of the panes, not a single
toolbar stretched across the full width of the container element), then
those toolbar items are pushed to the right side. Same goes for right ->
left.
- When there are vertically stacked panes, there is a divider that can
be dragged to resize. This uses Fluent's `useResizeHandle`. I tried to
use this for the existing width resizing, but ran into a bug, so I've
reached out to Fluent folks and am waiting on that.
- The top toolbar still combines top pane tabs and toolbar items. The
pane tabs for the lower panes now sit in the middle. The bottom toolbar
is still just toolbar items. I have some ongoing discussions with
Patrick about the final design for this, and it may change significantly
(e.g. we might switch to vertical tabs, or separate tabs from toolbar
items in some other way).
- When `embedMode` is enabled, this flows through to a `sidePaneMode` in
the `ShellService` (somewhat similar to the existing `toolbarMode`) that
pushes all left panes (top and bottom) to the top right quadrant and all
left panes (top and bottom) to the bottom right quadrant. This is a
generic heuristic that matches the behavior of Inspector v1.

As part of this change, I'm also updating all the Fluent libs to the
latest. This revealed a small issue where `ScrollToInterface` was being
imported from the old package location instead of the new one, which is
now fixed.

I also fixed a bug where perf stats needs to import
performanceViewerSceneExtension, otherwise there is a crash. I'm not
sure how this ever worked. Maybe something related to
performanceViewerSceneExtension changed more recently, not sure.

<img width="495" height="1334" alt="image"
src="https://github.com/user-attachments/assets/7f354737-b6e0-47a2-ae4d-dfe7eb7c123c"
/>