GUI 2D fix: properly restore context state after measure (#17213)
Control._applyStates() modifies global state on the context, and prior
to [this
commit](https://github.com/BabylonJS/Babylon.js/commit/20fd25bf114d6b357e26b255360d4289da21cb67),
context.save() was always called before calling _applyStates() and later
context.restore() called after the scenario's work was complete. That
change added a new callsite for _applyStates() which didn't call
save/restore, so it altered the context state unexpectedly.

With this fix, the new callsite calls save()/restore() correctly. Note
that save() and restore() operate on an internal stack, so nested calls
are OK.

To make it harder to make this mistake again, an eslint rule was added
as well as a docstring.

---------

Co-authored-by: AmoebaChant <kevbrown@microsoft.com>