gltfLoader: fully respect the skipMaterials flag (#17199)
There were two issues with `skipMaterials`:
1. If a mesh in the glTF has a material asigned but `skipMaterials` is
on, no material is loaded (correct behavior). If a mesh in the glTF has
no material assigned and `skipMaterials` is on, the flag is ignored and
a default material is still created and assigned.
2. `PBRMaterial` or `OpenPBRMaterial` are dynamically loaded, depending
on `useOpenPBR` flag. However, if `skipMaterials` is on, then we won't
be loading any materials anyway, so we should skip this dynamic import.

After these changes, I could see with some testing that when using
`skipMaterials` (with either a glb with materials and material
references and without), PBRMaterial was neither instantiated nor
imported (statically or dynamically).

Additionally, this change introduces the ability to completely disable
the default/PBR material loading even when `skipMaterials` is not
toggled on, which allows a loader extension to provide an alternative
material (such as a `NodeMaterial`).