Prevent WebGL error 'GL_INVALID_OPERATION: glDrawElements: Insufficient buffer size' (#17226)
When one mesh aggregates several distinct elements, one way to hide or
show a subset of these elements is too compute a new Index Buffer
depending on which element needs to turned On or Off.

But if all of them are off, this Index Buffer become empty.

Setting an empty Index Buffer to an Indexed Mesh results in a WebGL
error 'GL_INVALID_OPERATION: glDrawElements: Insufficient buffer size'
because internally the SubMesh that is created is not given `0`as
`indexCount`.

One workaround is to disabled the mesh is that case (that's I did in my
code), but nevertheless BabylonJs must not trigger WebGL errors.