Inspector v2: Various fixes and cleanup (#17232)
Sorry for the generic PR title... this is an assortment of fixes for
things I ran into while creating the Inspector v2 examples for the docs.

- Rename the test app's index.ts to index.tsx to make it easy to test
extensions that use React.
- Fix an issue where we had a couple instances of `return (entity as
AdvancedDynamicTexture)?.getClassName() === "AdvancedDynamicTexture";`
which will crash if the object does not have a `getClassName` function.
- Update the Pane helper component to be called SidePaneContainer (to
not conflict with Pane from shared-ui-components) and to make it fill
the space by default. Otherwise its not very useful.
- Fixed an issue where toolbar items added dynamically by extensions did
not show up immediately. This was just because of some missing
dependencies in a useMemo dependency list.
- Fixed a bunch of cases where we had `import { type foo } from "bar"`
instead of `import type { foo } from "bar"`.
- Auto ordered imports in files I touched to make it easier to see at a
glance what types are imported compared to real imports.