Viewer: Fix lighting regression now that we use PBRMaterial by default (#17233)
This PR fixes two Viewer regressions I noticed today:

1. For a model that does not have materials, a default PBRMaterial is
now used. However, some logic updates were missed that was causing a
default light to be created when it is not supposed to be, and an
environment not to be created when it is supposed to be. This made the
lighting different (and incorrect) depending on whether the model was
loaded after another model that does have materials (like a glb).
@alexandremottet
2. The camera was auto orbiting even if the flag is turned off. This
turned out to be an issue in the core behavior logic, where attaching a
behavior can be async/deferred, but in the meantime the behavior could
be removed. I'm not sure how this ever worked, but I think the change I
made here is more correct.