Fix floating origin bugs causing meshes to not be properly offset (#17234)
There were 2 bugs which were working together to hide themselves as
bugs.

2. The TempMatFinal variable was updating the asArray() function but the
internal _m was not getting updated. This change ensures the full matrix
is updated with the new float values.
3. Effect.setmatrix prototype override was directly calling
pipelineContext.setMatrix which had internal cache, so we were
mistakenly comparing offsetMatrix to cached regularMatrix. But because
of the above bug, the matrix sent to pipelineContext was actually not
offset, so the cache check would result in skipping the actual updating
of the uniform value. This change ensures the cache is checked on the
offset value

Tested the below playgrounds with floatingOrigin mode on, they all now
work
/#DLPNQT#0 - ragdoll
/#P7WF5X - physics glowlayer
/#4IUBBK#101 - simple single mesh

Co-authored-by: Georgina Halpern <gehalper@microsoft.com>