Inspector v2: New pane tab list (#17252)
This PR reworks the side pane tabs.

- Factored out a <Theme> component that wraps the Fluent ThemeProvider,
reads from local storage for the theme mode, and allows the theme to be
inverted.
- Used this <Theme> component for the main theme as well as inverted
themes for the pane header and the selected pane tab.
- Reduced the height of the toolbar (and therefore pane tabs).
- Changed the toolbar split buttons to not have the outline since it
looks awkward with the reduced height toolbar.
- Swapped out the pane tab implementation from TabList to Toolbar +
ToolbarRadioButton to give a look and feel that is a little closer to
Inspector v1.
- Reduced the height and font size for the pane header (to match the
toolbar).
- Fixed a bug where a useMemo dependency list was missing the pane tab
lists, which are now more dynamic with the theme inversion.

<img width="867" height="211" alt="image"
src="https://github.com/user-attachments/assets/0f389cc5-a400-4d8e-80d6-02dfda0c1869"
/>