Inspector v2: Fix various spacing issues in side panes (properties, stats, etc.) (#17259)
In a previous change we made the switch to have property lines generally
have a fixed height of 36px (per UX pass from @PatrickRyanMS). This was
initially implemented as:
- PropertyLine has a label on the left and some other control on the
right (like a switch or an input). The 36px fixed height was put here.
- PropertyLine contains a lower level LineContainer.
- LineContainer is used for other things, like ButtonLine, that don't
have a label but rather span the full width of the containing div.

The two main problems that I saw from this were:
1. Things using LineContainer (like ButtonLine) did not have a 36px
height, and did not have any padding (spacing between).
2. In situations where the natural height of the content exceeded 36px,
the content overflowed and obscured other lines.

To fix this, I made these changes:
1. Switch from a fixed `height` to a fixed `minHeight`.
2. Move the height constraint from PropertyLine to LineContainer.
3. Add a small amount of padding to the top/bottom of LineContainer (2px
each).
4. Switch the `boxSizing` to `border-box` so we can still specify the
minHeight for a line (36px) without having to do the math for any
padding and still end up with an overall 36px line height.

I also made a small change to the Textarea component to use the
`textarea` slot provided by the fluent control rather than having the
css style manually dig into the nested textarea. This is functionally
the same but cleaner and the supported mechanism provided by fluent.