Inspector v2: Various styling fixes (including some in shared UI) (#17263)
This PR fixes a bunch of styling issues across inspector v2 and shared
ui components, and also fixes a logic issue I discovered while fixing
the styling issues.

- Dropdown text is not truncated with ellipses rather than wrapping and
causing the dropdown size to grow vertically (often overflowing onto
other controls).
- Dropdown now applies the passed in css class to the wrapper div, not
to the nested fluent dropdown. This is needed so the passed in class can
adjust overall styling of the wrapper div (the whole control), like
setting the width.
- The 150px min width used in side panes (like the property pane) is now
applied in DropdownPropertyLine rather than in Dropdown itself. This
makes it possible to fix layout styling in the color picker since it
uses the Dropdown primitive.
- Fixed some other minor alignment issues in the color picker.
- Set the color picker ColorSwatch to use rounded corners to match the
rest of the styling.
- Fixed button alignment in TexturePreview so the buttons don't overflow
outside of the side pane.
- Moved a couple inline styles in Scene Explorer into classes (just got
missed in an earlier PR).

While fixing up the Dropdown issues, I also noticed a couple problems
with the sampling mode dropdown for textures:
1. The dropdown for the sampling mode property was targeting BaseTexture
instead of the lower level ThinTexture.
2. The sampling mode dropdown didn't actually work because it was bound
directly to the samplingMode property, which only has a getter. To
mutate it, you need to call `updateSamplingMode`. This is now fixed, but
I need to talk with @georginahalpern about having a compile time and/or
runtime check if you try to use BoundProperty on a readonly property.