Inspector v2: Various post announce fixes (#17282)
A few Inspector v2 issues surfaced after the preview announcement which
this PR addresses:
- There is a crash when selecting various entities in scene explorer,
which is caused by passing a negative value to `toFixed` (a negative
value returned by `CalculatePrecision`). This is fixed by making sure
the precision we pass to `toFixed` is a minimum of 0.
- Fix a label typo (Indeces -> Indices).
- Add a setting for automatically selecting/showing the properties pane
when an entity is selected (on by default).
- Update the feedback button to link directly to the forum announcement
thread.