Disallow Inspector-v2 in Playground/Sandbox when old Babylon versions are specified (#17296)
This PR fixes an issue I missed where if in Playground/Sandbox you use
an older version of Babylon, things can break. This is because we don't
have old versions of inspector v2 that work with old versions of Babylon
like we do with inspector v1 (plus we don't have a umd package right
now, so we don't even have a means of loading older versions of
inspector v2).

For now with this change, we only allow Inspector v2 when using the
latest Babylon (e.g. no explicit version specified).