Gltf: unofficial _COLOR_0 for EXT_mesh_gpu_instancing (#17264)
**Loader**

Added support of loading `_COLOR_0` attribute to thin instance color,
both RGB and RGBA is supported and would be loaded as-is, since it's
supported [at the rendering context
level](https://forum.babylonjs.com/t/gltf-unofficial-color-0-for-ext-mesh-gpu-instancing/60455/10).

**Serializer**

Added support of serializing thin instance color to `_COLOR_0`
attribute, this is disabled by default to keep behavior compatible with
previous versions, but can be enabled with a new experimental option
`exportInstanceColor` of the serializer.
Since babylon.js uses RGBA for thin instance color by default, and the
widely used unofficial `_COLOR_0` attribute was RGB, serializer defaults
to converting RGBA to RGB, and emit a warning when instance color alpha
enabled via `hasVertexAlpha`. This behavior can be configured with a
newly added experimental option of the serializer.

**Forum post**

<https://forum.babylonjs.com/t/gltf-unofficial-color-0-for-ext-mesh-gpu-instancing/60455>