Fix KHR_mesh_visibility issue with animation pointers (#17285)
This change fixes an issue with the [sample asset for
KHR_mesh_visibility](https://github.com/KhronosGroup/glTF-Sample-Assets/pull/235)
not working. I'm not adding a test for this asset yet as this sample
asset is not conformant with the extension spec yet.

It also moves the visible properties from `AbstractMesh` to `Node` which
fixes some issues with how visibility is being handled. The inspector v2
has also been updated to match this change.