Inspector v2: Fix Scene Explorer max depth of 10 (#17304)
This PR has a few fixes for Inspector v2, mostly feedback from the
forum:
1. Scene Explorer wonked out once it hit a node that was 10 levels or
deeper. Interestingly this is a
[documented](https://storybooks.fluentui.dev/react/?path=/docs/components-tree--docs#inline-styling-tree-item-level)
as being the default for performance reasons, but there is a way to fix
it. I did not notice any impact on perf.
2. Adding a `debugNode` global like Inspector v1 (see
https://forum.babylonjs.com/t/scene-debuglayer-copy-last-command-to-clipboard/49796/18
for context).
3. Fix an issue I noticed where there was a missing useCallback in the
recent "mini stats" I added that was causing a lot of unnecessary react
renders and frequent console errors as a result.